+++ src/app/index.ts
@@ -1,4 +1,3 @@
-export * from './appMode'
 export * from './appVersion'
 export * from './browserProjectRuntime'
 export * from './projectRuntime'